+++ release-notes.md
@@ -1,4 +1,9 @@
 Release notes
 -------------
+##### v1.0.1 (2013-05-20)
+New features and bugfixes  
+`+` square rotation is now optional: `behavior.rotate`  
+`-` fixed jagged edges in Firefox  
+
 ##### v1.0.0 (2013-05-15)
 Initial commit